Rearrange some FS client functions
