+++ src/LibHac/Fs/ICommonMountNameGenerator.cs
@@ -4,6 +4,6 @@ namespace LibHac.Fs
 {
     public interface ICommonMountNameGenerator
     {
-        Result Generate(Span<byte> nameBuffer);
+        Result GenerateCommonMountName(Span<byte> nameBuffer);
     }
 }
