+++ src/LibHac/Fs/Shim/RightsId.cs
@@ -3,12 +3,13 @@ using LibHac.FsService;
 using LibHac.FsSystem;
 using LibHac.Ncm;
 using LibHac.Spl;
+using FsRightsId = LibHac.Fs.RightsId;
 
-namespace LibHac.Fs
+namespace LibHac.Fs.Shim
 {
-    public static class ExternalKeys
+    public static class RightsId
     {
-        public static Result GetRightsId(this FileSystemClient fs, out RightsId rightsId, TitleId programId,
+        public static Result GetRightsId(this FileSystemClient fs, out FsRightsId rightsId, TitleId programId,
             StorageId storageId)
         {
             IFileSystemProxy fsProxy = fs.GetFileSystemProxyServiceObject();
@@ -16,7 +17,7 @@ namespace LibHac.Fs
             return fsProxy.GetRightsId(out rightsId, programId, storageId);
         }
 
-        public static Result GetRightsId(this FileSystemClient fs, out RightsId rightsId, U8Span path)
+        public static Result GetRightsId(this FileSystemClient fs, out FsRightsId rightsId, U8Span path)
         {
             rightsId = default;
 
@@ -28,7 +29,7 @@ namespace LibHac.Fs
             return fsProxy.GetRightsIdByPath(out rightsId, ref fsPath);
         }
 
-        public static Result GetRightsId(this FileSystemClient fs, out RightsId rightsId, out byte keyGeneration, U8Span path)
+        public static Result GetRightsId(this FileSystemClient fs, out FsRightsId rightsId, out byte keyGeneration, U8Span path)
         {
             rightsId = default;
             keyGeneration = default;
@@ -41,14 +42,14 @@ namespace LibHac.Fs
             return fsProxy.GetRightsIdAndKeyGenerationByPath(out rightsId, out keyGeneration, ref fsPath);
         }
 
-        public static Result RegisterExternalKey(this FileSystemClient fs, ref RightsId rightsId, ref AccessKey key)
+        public static Result RegisterExternalKey(this FileSystemClient fs, ref FsRightsId rightsId, ref AccessKey key)
         {
             IFileSystemProxy fsProxy = fs.GetFileSystemProxyServiceObject();
 
             return fsProxy.RegisterExternalKey(ref rightsId, ref key);
         }
 
-        public static Result UnregisterExternalKey(this FileSystemClient fs, ref RightsId rightsId)
+        public static Result UnregisterExternalKey(this FileSystemClient fs, ref FsRightsId rightsId)
         {
             IFileSystemProxy fsProxy = fs.GetFileSystemProxyServiceObject();
